+++ src/router/metadata/metadataHandler.test.ts
@@ -3,19 +3,21 @@
  *  SPDX-License-Identifier: Apache-2.0
  */
 
-import { stubs } from 'fhir-works-on-aws-interface';
+import { stubs, Persistence } from 'fhir-works-on-aws-interface';
+import each from 'jest-each';
 import MetadataHandler from './metadataHandler';
 import { makeOperation } from './cap.rest.resource.template';
 import r4FhirConfigGeneric from '../../../sampleData/r4FhirConfigGeneric';
 import r4FhirConfigWithExclusions from '../../../sampleData/r4FhirConfigWithExclusions';
 import stu3FhirConfigWithExclusions from '../../../sampleData/stu3FhirConfigWithExclusions';
 import r4FhirConfigNoGeneric from '../../../sampleData/r4FhirConfigNoGeneric';
-import Validator from '../validation/validator';
+import JsonSchemaValidator from '../validation/jsonSchemaValidator';
 import ConfigHandler from '../../configHandler';
 import { utcTimeRegExp } from '../../regExpressions';
+import { FHIRStructureDefinitionRegistry } from '../../registry';
 
-const r4Validator = new Validator('4.0.1');
-const stu3Validator = new Validator('3.0.1');
+const r4Validator = new JsonSchemaValidator('4.0.1');
+const stu3Validator = new JsonSchemaValidator('3.0.1');
 
 const SUPPORTED_R4_RESOURCES = [
     'Account',
@@ -327,6 +329,7 @@ const overrideStubs = {
         }),
     },
 };
+const registry: FHIRStructureDefinitionRegistry = new FHIRStructureDefinitionRegistry();
 
 describe('ERROR: test cases', () => {
     beforeEach(() => {
@@ -339,7 +342,7 @@ describe('ERROR: test cases', () => {
             stu3FhirConfigWithExclusions(),
             SUPPORTED_STU3_RESOURCES,
         );
-        const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+        const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
         try {
             // OPERATE
             await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
@@ -357,7 +360,7 @@ describe('ERROR: test cases', () => {
             r4FhirConfigGeneric(overrideStubs),
             SUPPORTED_R4_RESOURCES,
         );
-        const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+        const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
         try {
             // OPERATE
             await metadataHandler.capabilities({ fhirVersion: '3.0.1', mode: 'full' });
@@ -374,7 +377,7 @@ test('STU3: FHIR Config V3 with 2 exclusions and search', async () => {
     const config = stu3FhirConfigWithExclusions(overrideStubs);
     const supportedGenericResources = ['AllergyIntolerance', 'Organization', 'Account', 'Patient'];
     const configHandler: ConfigHandler = new ConfigHandler(config, supportedGenericResources);
-    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, true);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry, true);
     const response = await metadataHandler.capabilities({ fhirVersion: '3.0.1', mode: 'full' });
     const { genericResource } = config.profile;
     const excludedResources = genericResource ? genericResource.excludedSTU3Resources || [] : [];
@@ -395,7 +398,7 @@ test('STU3: FHIR Config V3 with 2 exclusions and search', async () => {
         }
         const expectedResourceSubset = {
             interaction: makeOperation(['read', 'create', 'update', 'vread', 'search-type']),
-            updateCreate: true,
+            updateCreate: configHandler.config.profile.genericResource!.persistence.updateCreateSupported,
             searchParam: [
                 {
                     name: 'some-search-field',
@@ -410,13 +413,11 @@ test('STU3: FHIR Config V3 with 2 exclusions and search', async () => {
 
     expect(response.resource.rest[0].interaction).toEqual(makeOperation(config.profile.systemOperations));
     expect(response.resource.rest[0].searchParam).toBeUndefined();
-    expect(stu3Validator.validate('CapabilityStatement', response.resource)).toEqual({
-        message: 'Success',
-    });
+    expect(stu3Validator.validate(response.resource)).resolves.toEqual(undefined);
 });
 test('R4: FHIR Config V4 without search', async () => {
     const configHandler: ConfigHandler = new ConfigHandler(r4FhirConfigGeneric(overrideStubs), SUPPORTED_R4_RESOURCES);
-    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
     const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
     expect(response.resource).toBeDefined();
     expect(response.resource.acceptUnknown).toBeUndefined();
@@ -427,22 +428,20 @@ test('R4: FHIR Config V4 without search', async () => {
     // see if the four CRUD + vRead operations are chosen
     const expectedResourceSubset = {
         interaction: makeOperation(['create', 'read', 'update', 'delete', 'vread', 'history-instance']),
-        updateCreate: true,
+        updateCreate: configHandler.config.profile.genericResource!.persistence.updateCreateSupported,
     };
     expect(response.resource.rest[0].resource[0]).toMatchObject(expectedResourceSubset);
     expect(response.resource.rest[0].interaction).toEqual(
         makeOperation(r4FhirConfigGeneric(overrideStubs).profile.systemOperations),
     );
     expect(response.resource.rest[0].searchParam).toBeUndefined();
-    expect(r4Validator.validate('CapabilityStatement', response.resource)).toEqual({
-        message: 'Success',
-    });
+    expect(r4Validator.validate(response.resource)).resolves.toEqual(undefined);
 });
 
 test('R4: FHIR Config V4 with 3 exclusions and AllergyIntollerance special', async () => {
     const config = r4FhirConfigWithExclusions(overrideStubs);
     const configHandler: ConfigHandler = new ConfigHandler(config, SUPPORTED_R4_RESOURCES);
-    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
     const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
     const { genericResource } = config.profile;
     const excludedResources = genericResource ? genericResource.excludedR4Resources || [] : [];
@@ -464,12 +463,12 @@ test('R4: FHIR Config V4 with 3 exclusions and AllergyIntollerance special', asy
         if (resource.type === 'AllergyIntolerance') {
             expectedResourceSubset = {
                 interaction: makeOperation(['create', 'update']),
-                updateCreate: true,
+                updateCreate: configHandler.config.profile.genericResource!.persistence.updateCreateSupported,
             };
         } else {
             expectedResourceSubset = {
                 interaction: makeOperation(['read', 'history-instance', 'history-type']),
-                updateCreate: false,
+                updateCreate: configHandler.config.profile.genericResource!.persistence.updateCreateSupported,
             };
         }
         expect(resource).toMatchObject(expectedResourceSubset);
@@ -478,15 +477,13 @@ test('R4: FHIR Config V4 with 3 exclusions and AllergyIntollerance special', asy
     expect(isExclusionFound).toBeFalsy();
     expect(response.resource.rest[0].interaction).toEqual(makeOperation(config.profile.systemOperations));
     expect(response.resource.rest[0].searchParam).toBeDefined();
-    expect(r4Validator.validate('CapabilityStatement', response.resource)).toEqual({
-        message: 'Success',
-    });
+    expect(r4Validator.validate(response.resource)).resolves.toEqual(undefined);
 });
 
 test('R4: FHIR Config V4 no generic set-up & mix of STU3 & R4', async () => {
     const config = r4FhirConfigNoGeneric(overrideStubs);
     const configHandler: ConfigHandler = new ConfigHandler(config, SUPPORTED_R4_RESOURCES);
-    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
     const configResource: any = config.profile.resources;
     const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
     expect(response.resource).toBeDefined();
@@ -503,7 +500,7 @@ test('R4: FHIR Config V4 no generic set-up & mix of STU3 & R4', async () => {
         }
         const expectedResourceSubset = {
             interaction: makeOperation(configResource[resource.type].operations),
-            updateCreate: configResource[resource.type].operations.includes('update'),
+            updateCreate: configHandler.config.profile.resources![resource.type].persistence.updateCreateSupported,
         };
         expect(resource).toMatchObject(expectedResourceSubset);
         if (configResource[resource.type].operations.includes('search-type')) {
@@ -517,15 +514,38 @@ test('R4: FHIR Config V4 no generic set-up & mix of STU3 & R4', async () => {
         makeOperation(r4FhirConfigNoGeneric().profile.systemOperations),
     );
     expect(response.resource.rest[0].searchParam).toBeDefined();
-    expect(r4Validator.validate('CapabilityStatement', response.resource)).toEqual({
-        message: 'Success',
+    expect(r4Validator.validate(response.resource)).resolves.toEqual(undefined);
+});
+
+each([
+    ['Generic Resources: updateCreate = true', true, r4FhirConfigGeneric],
+    ['Generic Resources: updateCreate = false', false, r4FhirConfigGeneric],
+    ['Special Resources: updateCreate = true', true, r4FhirConfigNoGeneric],
+    ['Special Resources: updateCreate = false', false, r4FhirConfigNoGeneric],
+]).test('R4: FHIR Config with %s', async (testName: string, updateCreateSupported: boolean, fhirConfigBuilder: any) => {
+    const persistence: Persistence = {
+        ...stubs.persistence,
+        updateCreateSupported,
+    };
+
+    const fhirConfig = fhirConfigBuilder({ persistence, ...overrideStubs });
+
+    const configHandler: ConfigHandler = new ConfigHandler(fhirConfig, SUPPORTED_R4_RESOURCES);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
+    const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
+    response.resource.rest[0].resource.forEach((resource: any) => {
+        const expectedResourceSubset = {
+            updateCreate: updateCreateSupported,
+        };
+        expect(resource).toMatchObject(expectedResourceSubset);
     });
 });
+
 test('R4: FHIR Config V4 with bulkDataAccess', async () => {
     const r4ConfigWithBulkDataAccess = r4FhirConfigGeneric(overrideStubs);
     r4ConfigWithBulkDataAccess.profile.bulkDataAccess = stubs.bulkDataAccess;
     const configHandler: ConfigHandler = new ConfigHandler(r4ConfigWithBulkDataAccess, SUPPORTED_R4_RESOURCES);
-    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
     const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
 
     expect(response.resource.rest[0].operation).toEqual([
@@ -544,7 +564,7 @@ test('R4: FHIR Config V4 with bulkDataAccess', async () => {
 
 test('R4: FHIR Config V4 without bulkDataAccess', async () => {
     const configHandler: ConfigHandler = new ConfigHandler(r4FhirConfigGeneric(overrideStubs), SUPPORTED_R4_RESOURCES);
-    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
     const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
 
     expect(response.resource.rest[0].operation).toBeUndefined();
@@ -564,7 +584,7 @@ test('R4: FHIR Config V4 with all Oauth Policy endpoints', async () => {
         },
     };
     const configHandler: ConfigHandler = new ConfigHandler(r4ConfigWithOauthEndpoints, SUPPORTED_R4_RESOURCES);
-    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
     const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
 
     expect(response.resource.rest[0].security).toEqual({
@@ -606,7 +626,7 @@ test('R4: FHIR Config V4 with all Oauth Policy endpoints', async () => {
                 coding: [
                     {
                         code: 'OAuth',
-                        system: 'https://www.hl7.org/fhir/codesystem-restful-security-service.html',
+                        system: 'http://hl7.org/fhir/ValueSet/restful-security-service',
                     },
                 ],
             },
@@ -625,7 +645,7 @@ test('R4: FHIR Config V4 with some Oauth Policy endpoints', async () => {
         },
     };
     const configHandler: ConfigHandler = new ConfigHandler(r4ConfigWithOauthEndpoints, SUPPORTED_R4_RESOURCES);
-    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
     const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
 
     expect(response.resource.rest[0].security).toEqual({
@@ -655,7 +675,7 @@ test('R4: FHIR Config V4 with some Oauth Policy endpoints', async () => {
                 coding: [
                     {
                         code: 'OAuth',
-                        system: 'https://www.hl7.org/fhir/codesystem-restful-security-service.html',
+                        system: 'http://hl7.org/fhir/ValueSet/restful-security-service',
                     },
                 ],
             },
@@ -675,7 +695,7 @@ test('R4: FHIR Config V4 with all productInfo params', async () => {
         copyright: 'Copyright',
     };
     const configHandler: ConfigHandler = new ConfigHandler(r4ConfigWithAllProductInfo, SUPPORTED_R4_RESOURCES);
-    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, registry);
     const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
 
     const expectedResponse: any = {