+++ src/router/metadata/metadataHandler.ts
@@ -10,26 +10,33 @@ import makeSecurity from './cap.rest.security.template';
 import makeRest from './cap.rest.template';
 import makeStatement from './cap.template';
 import ConfigHandler from '../../configHandler';
+import { FHIRStructureDefinitionRegistry } from '../../registry';
 
 export default class MetadataHandler implements Capabilities {
     configHandler: ConfigHandler;
 
     readonly hasCORSEnabled: boolean;
 
-    constructor(handler: ConfigHandler, hasCORSEnabled: boolean = false) {
+    readonly registry: FHIRStructureDefinitionRegistry;
+
+    constructor(handler: ConfigHandler, registry: FHIRStructureDefinitionRegistry, hasCORSEnabled: boolean = false) {
         this.configHandler = handler;
         this.hasCORSEnabled = hasCORSEnabled;
+        this.registry = registry;
     }
 
     private async generateResources(fhirVersion: FhirVersion) {
         const specialResourceTypes = this.configHandler.getSpecialResourceTypes(fhirVersion);
         let generatedResources = [];
         if (this.configHandler.config.profile.genericResource) {
+            const updateCreate = this.configHandler.config.profile.genericResource.persistence.updateCreateSupported;
             const generatedResourcesTypes = this.configHandler.getGenericResources(fhirVersion, specialResourceTypes);
             generatedResources = makeGenericResources(
                 generatedResourcesTypes,
                 this.configHandler.getGenericOperations(fhirVersion),
                 await this.configHandler.config.profile.genericResource.typeSearch.getCapabilities(),
+                await this.registry.getCapabilities(),
+                updateCreate,
             );
         }